% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCnormalization.R
\name{GCnormalization}
\alias{GCnormalization}
\title{GC normalization
This function normalizes the coverage values and limits the GC content bias induced by PCR.
Inputs: coverage dataframe (1st col POS, 2nd col CCOVERAGE), path to fasta file
Output:  vector of modified COVERAGE values}
\usage{
GCnormalization(coverage, fastaFile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A vector of new coverage values
}
\description{
GC normalization
This function normalizes the coverage values and limits the GC content bias induced by PCR.
Inputs: coverage dataframe (1st col POS, 2nd col CCOVERAGE), path to fasta file
Output:  vector of modified COVERAGE values
}
