% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesd.R
\name{gesd}
\alias{gesd}
\title{GESD - Two-sided Generalized Extreme Studentized Deviate test
This function requires:
1 - a dataframe with 2 columns - 1st column: position/coordinate 2nd column: coverage/read-depth
2- an upper bound for the suspected number of outliers (an integer)
It returns outliers indices as vector
See https://www.itl.nist.gov/div898/handbook/eda/section3/eda35h3.htm}
\usage{
gesd(coverageDF, estimatedOutliers)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A vector of outliers indices from input
}
\description{
GESD - Two-sided Generalized Extreme Studentized Deviate test
This function requires:
1 - a dataframe with 2 columns - 1st column: position/coordinate 2nd column: coverage/read-depth
2- an upper bound for the suspected number of outliers (an integer)
It returns outliers indices as vector
See https://www.itl.nist.gov/div898/handbook/eda/section3/eda35h3.htm
}
