% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAPPABILITYnormalization.R
\name{MAPPABILITYnormalization}
\alias{MAPPABILITYnormalization}
\title{MAPPABILITY normalization
This function normalizes the coverage values and limits the GC content bias induced by PCR.
Inputs: coverage dataframe (1st col POS, 2nd col CCOVERAGE), path to bedgraph file
Output: vector of modified COVERAGE values}
\usage{
MAPPABILITYnormalization(coverage, bedgraphFile)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A vector of new coverage values
}
\description{
MAPPABILITY normalization
This function normalizes the coverage values and limits the GC content bias induced by PCR.
Inputs: coverage dataframe (1st col POS, 2nd col CCOVERAGE), path to bedgraph file
Output: vector of modified COVERAGE values
}
