% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzscore.R
\name{mzscore}
\alias{mzscore}
\title{Modified Z-score by Iglewicz and Hoaglin
This function loads a vector of coverage/read-depth values. It returns the indices of potential outliers.
See https://www.itl.nist.gov/div898/handbook/eda/section3/eda35h.htm}
\usage{
mzscore(x)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A vector of outliers indices from input
}
\description{
Modified Z-score by Iglewicz and Hoaglin
This function loads a vector of coverage/read-depth values. It returns the indices of potential outliers.
See https://www.itl.nist.gov/div898/handbook/eda/section3/eda35h.htm
}
