% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ORICnormalization.R
\name{ORICnormalization}
\alias{ORICnormalization}
\title{REPLICATION ORIGIN BIAS normalization
This function normalizes the replication origin bias.
Inputs: coverage dataframe (1st col POS, 2nd col CCOVERAGE), position of replication origin in reference
Output: vector of modified COVERAGE values}
\usage{
ORICnormalization(coverage, oriC_position = 1)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A vector of new coverage values
}
\description{
REPLICATION ORIGIN BIAS normalization
This function normalizes the replication origin bias.
Inputs: coverage dataframe (1st col POS, 2nd col CCOVERAGE), position of replication origin in reference
Output: vector of modified COVERAGE values
}
