% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVCF.R
\name{writeVCF}
\alias{writeVCF}
\title{Write VCF file from CNproScan dataframe
Function to reformat CNV dataframe to VCF formated dataframe to write down
Inputs: detected CNV dataframe
Output: VCF file}
\usage{
writeVCF(CNV_DF, fileName)
}
\arguments{
\item{CNV_DF}{CNV dataframe}

\item{fileName}{name of the VCF file}
}
\value{
VCF dataframe ready to write to disk
}
\description{
Write VCF file from CNproScan dataframe
Function to reformat CNV dataframe to VCF formated dataframe to write down
Inputs: detected CNV dataframe
Output: VCF file
}
