% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNproScan.R
\name{CNproScanCNV}
\alias{CNproScanCNV}
\title{CNproScan - CNV detection for bacteria
This function detects and annotate CNVs in bacterial genomes. It uses GESD outliers detection
and detection of discordant read-pairs for annotation.
Last Update: 5/10/2021}
\usage{
CNproScanCNV(coveragefile, bamFile, fastaFile, cores = 2)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A vector of outliers indices from input
}
\description{
CNproScan - CNV detection for bacteria
This function detects and annotate CNVs in bacterial genomes. It uses GESD outliers detection
and detection of discordant read-pairs for annotation.
Last Update: 5/10/2021
}
