% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeVCF.R
\name{writeVCF}
\alias{writeVCF}
\title{Write VCF file
Function to reformat CNV dataframe to VCF formated dataframe to write down
Inputs: detected CNV dataframe
Output: VCF file}
\usage{
writeVCF(CNV_DF, chromosome_name)
}
\arguments{
\item{infile}{CNV dataframe}
}
\value{
VCF dataframe ready to write to disk
}
\description{
Write VCF file
Function to reformat CNV dataframe to VCF formated dataframe to write down
Inputs: detected CNV dataframe
Output: VCF file
}
